feat: UserService 李먮쭑..?

+++ src/backend/user-service/UserService/Business/Friend/service/FriendService.cs
@@ -1,10 +1,11 @@
+using System.Text;
 using Castle.DynamicProxy;
-using user_service.common;
+using Newtonsoft.Json;
+using user_service.Business.Friend.dto;
 using user_service.common.exception;
 using user_service.Controllers.dto.friend;
 using user_service.friend.repository;
 using user_service.intercepter;
-using user_service.user.dto;
 
 namespace user_service
 {
@@ -15,31 +16,75 @@ namespace user_service
             public class FriendService : IFriendService
             {
                 private IConfiguration _config;
-                private IUserRepository _userRepository;
                 private IFriendRepository _friendRepository;
-
+                private HttpClient _httpClient;
                 public FriendService(IConfiguration config,
-                                    IUserRepository userRepository,
                                     IFriendRepository friendRepository,
                                     LogInterceptor interceptor)
                 {
                     var generator = new ProxyGenerator();
-                    _userRepository = generator.CreateInterfaceProxyWithTargetInterface<IUserRepository>(userRepository, interceptor);
+                    _friendRepository = generator.CreateInterfaceProxyWithTargetInterface<IFriendRepository>(friendRepository, interceptor);
                     _config = config;
-                    _friendRepository = friendRepository;
+                    _httpClient = new HttpClient();
+                    _httpClient.BaseAddress = new Uri("http://10.99.14.176:9090/api/state/direct/user/info");
                 }
 
-                public List<UserDTO> GetFriendList(long id)
+                public async Task<List<FriendInfoDTO>> GetFriendList(long id)
                 {
-                    return _friendRepository.GetFriendList(id);
+                    List<FriendInfoDTO> friendInfos = _friendRepository.GetFriendList(id);
+
+                    UserIdsDTO userIds = new();
+                    foreach (FriendInfoDTO friendInfo in friendInfos)
+                    {
+                        userIds.userIds.Add(friendInfo.Id);
+                    }
+
+                    // 상태관리 서버에게 전송해야함
+                    using StringContent jsonContent = new(
+                        JsonConvert.SerializeObject(userIds),
+                        Encoding.UTF8,
+                        "application/json");
+
+                    CancellationTokenSource cts = new CancellationTokenSource();
+                    cts.CancelAfter(100); // Cancel after 1 second
+                    CancellationToken cancellationToken = cts.Token;
+                    try
+                    {
+                        var response = await _httpClient.PostAsync("", jsonContent, cancellationToken);
+                        if (response.StatusCode == System.Net.HttpStatusCode.OK)
+                        {
+                            var result = await response.Content.ReadAsStringAsync();
+                            ConnectsStatusDTO? data = JsonConvert.DeserializeObject<ConnectsStatusDTO>(result);
+                            if (data == null)
+                                throw new ServiceException(4025);
+
+                            foreach (FriendInfoDTO friendInfo in friendInfos)
+                            {
+                                if (data.connectionStates[friendInfo.Id.ToString()] == "online")
+                                    friendInfo.IsOnline = true;
+                                else
+                                    friendInfo.IsOnline = false;
+                            }
+                        }
+                        else
+                        {
+                            throw new ServiceException(4025);
+                        }
+                    }
+                    catch (System.Exception)
+                    {
+
+                    }
+
+                    return friendInfos;
                 }
 
-                public List<UserDTO> GetSendRequestList(long id)
+                public List<FriendInfoDTO> GetSendRequestList(long id)
                 {
                     return _friendRepository.ShowAllSendRequestList(id);
                 }
 
-                public List<UserDTO> GetReceiveRequestList(long id)
+                public List<FriendInfoDTO> GetReceiveRequestList(long id)
                 {
                     return _friendRepository.ShowAllReceiveRequesttList(id);
                 }
@@ -49,11 +94,14 @@ namespace user_service
                     long id = friend.MyId;
                     long friendId = GetFriendId(friend.FriendEmail);
 
-                    if(_friendRepository.CheckAlreadyFriend(id, friendId))
+                    if (id == friendId)
+                        throw new ServiceException(4026);
+
+                    if (_friendRepository.CheckAlreadyFriend(id, friendId))
                         throw new ServiceException(4016);
 
-                    if(!_friendRepository.SendFriendRequest(id, friendId))
-                        throw new ServiceException(4017); 
+                    if (!_friendRepository.SendFriendRequest(id, friendId))
+                        throw new ServiceException(4017);
                 }
 
                 public void CancleFriendAddRequest(FriendDTO friend)
@@ -61,7 +109,10 @@ namespace user_service
                     long id = friend.MyId;
                     long friendId = GetFriendId(friend.FriendEmail);
 
-                    if(!_friendRepository.CancleFriendRequest(id, friendId))
+                    if (id == friendId)
+                        throw new ServiceException(4026);
+
+                    if (!_friendRepository.CancleFriendRequest(id, friendId))
                         throw new ServiceException(4020);
                 }
 
@@ -69,8 +120,11 @@ namespace user_service
                 {
                     long id = friend.MyId;
                     long friendId = GetFriendId(friend.FriendEmail);
-                    
-                    if(!_friendRepository.AcceptFriendRequest(id, friendId))
+
+                    if (id == friendId)
+                        throw new ServiceException(4026);
+
+                    if (!_friendRepository.AcceptFriendRequest(id, friendId))
                         throw new ServiceException(4018);
                 }
 
@@ -79,7 +133,10 @@ namespace user_service
                     long id = friend.MyId;
                     long friendId = GetFriendId(friend.FriendEmail);
 
-                    if(!_friendRepository.RefuseFriendRequest(id, friendId))
+                    if (id == friendId)
+                        throw new ServiceException(4026);
+
+                    if (!_friendRepository.RefuseFriendRequest(id, friendId))
                         throw new ServiceException(4019);
                 }
 
@@ -88,14 +145,18 @@ namespace user_service
                     long id = friend.MyId;
                     long friendId = GetFriendId(friend.FriendEmail);
 
+                    if (id == friendId)
+                        throw new ServiceException(4027);
+
                     return _friendRepository.DeleteFriend(id, friendId);
                 }
 
                 private long GetFriendId(string email)
                 {
                     long friendId = _friendRepository.GetFriendId(email);
-                    if(friendId == 0)
+                    if (friendId == 0)
                         throw new ServiceException(4007);
+
                     return friendId;
                 }
             }
